Updating html and css with new content.
Had to add class to <a> tags in main nav, as I used the <a> tag in other places and didn't want it to be styled the same way as main nav.
Added "back to top" option so it's quicker to get back to the beginning of the page. 
Started styling mid-size view-port.